Add the base structure for the Slashers game
This commit introduces the foundational structure for the Slashers game. It adds new files for the game's main logic, the Pawn entity, and the Pawn's controller. A rough UI for the game setup and the HUD are implemented via the Razor syntax. Some player properties, like input direction and view angles, are now networked to keep things in-sync. Weapons are more modular now, introducing a new Weapon class and a ViewModel for it. A `.gitignore` file is added to ignore common directories and files not needed in the Git repository, a README file is added to provide information about the project, and a `.sbproj` file is added to define the s&box project. The `.editorconfig` file is also included to maintain coding style consistencies across different editors and IDEs used by collaborators. A loading screen and the asset files for UI elements are included as well.