Add GameFlowModule and GameModule
Added two new files, GameFlowModule.cs and GameModule.cs, in the Manager directory to handle activation and deactivation of the game module. Also made some modifications in Game.cs file, including the addition of necessary namespaces, renaming the game class and making it a Prefab. Improved the game's architecture to segregate responsibilities and increase modularity.