Refactor input handling in camera and movement controllers
Removed the invocation of BuildInput() from CameraController and reintroduced it into MovementsController. This change of responsibility aligns better with each controller's area of control and ensures a better organized and readable codebase. The Pawn class has been simplified by removing it's unnecessary implementation of BuildInput() as a result of this transition.