Add footstep sound logic to Pawn.cs
Added a new method `OnAnimEventFootstep` to implement footstep sound logic in Pawn.cs. The logic includes a time limit between the footstep sounds, creating a realism in the footstep sound production. We're achieving that by keeping timestamp of last footstep sound played. When it receives a footstep event it checks if the time elapsed from last footstep is less than 0.2 sec, if so it ignores the event. This prevents the excessive and unrealistic footstep sounds.