Add stuck check, shift event handling to EventDispatcher
Implemented a new method, CheckStuckAndFix, within the MovementController to validate and handle instances when a character is stuck. This is to improve the gaming experience by providing a solution to the stuck character problem.
Also, event handling in AnimationController and MovementController has been moved from internal HashSet structure to an external EventDispatcher instance attached to the Pawn instance. This shifts management of events to the EventDispatcher, leading to cleaner, more manageable code and potentially improving the performance.
Also added EventOnJump and EventOnGrounded events that are dispatched from MovementsController when the character jumps or lands.