Refactor event unregistering method for improved clarity and efficiency
Altered UnregisterAllEvents method to improve code clarity and efficiency. Firstly, the ArgumentNullException case was removed to simplify the code, the functionality of the code was not affected by this removal. Instead of directly modifying the original dictionary while it's being iterated, a new list (keysToRemove) was added to store keys for removal. Doing so could prevent possible exceptions due to concurrent modification and improve code safety. Also, callback.Invoke was called directly regardless of its type, reducing unnecessary type checking for Action<Event> and speeding up the event delivery process.