"Enhance movements and debug info in the game
The commit introduces walking and sprinting speeds to MovementsController.cs along with a corresponding state transition system for movements in MovementsController.HFSM.cs. A CheckStuckAndFix method is introduced in pawn movements to ensure that the player does not get stuck while moving. The change in the state transition system also includes a new 'OnEnterWalkState' method to handle entry into the walking state. Correspondingly, the respective speeds for walking and sprinting have been set in MovementsController.OnGroundState.cs.

Debug logging in Utils.HFSM.cs has been enabled and the components are now created on creation of Pawn in Pawn.cs instead of when it respawns for improving debug info.

In s&box projects, particularly in the slashers game, the player's movement functionality now includes walking and sprinting. The debug information has also been enhanced for better issue tracking and resolution."