Add ducking functionality to game character
Added methods in MovementsController.cs to allow the character to duck and jump if grounded. Updated AnimatorController.cs to trigger jump and duck animations. Created a state machine in MovementsController.HFSM.cs to handle transitions related to ducking, walking, running and sprinting. Implemented ducking and unducking logic in MovementsController.OnGroundState.cs with corresponding events and added a Coroutine enumerator to smoothly resize the character's hull when ducking/unducking. This update is intended to provide more realistic character movements in the game.