Add ducking transition from in-air state
A ducking transition from the in-air state to the duck state was added. This change allows players to initiate a duck action while they are in the air, increasing flexibility in movement mechanics. The 'duck' input and 'grounded' check has been used for deciding the transition. Modified the DuckState handling, added halt to an existing coroutine before starting the ducking routine.