Refactored HFSM implementation to enable method chaining
Changed the HFSM methods in Utils/HFSM.cs to include method chaining. This makes the implementation more robust and easier to use, which leads to cleaner and more readable client code. Each method that previously returned void, now returns an instance of HFSMBuilder, allowing the calls to be chained together. This change does not affect the overall functionality of the system.