Add Knife weapon and adjust Player Respawn event
Implement a new Knife weapon into the game and adjust how the weapons are added during Player Respawn event. The Pistol is no long the active weapon upon respawn, instead the Knife is now the active weapon.

The Knife class has been created with all the respective methods and attributes such as AttackPrimary and AttackSecondary. Some properties of the Weapon class were adjusted to accommodate for the Knife, such as the Handedness property. Changes in the Pistol class were also made in order to keep the code cleaner.