Add swimming state and related events to movement controller
A new swimming state has been introduced to the game's movement controller. This state is triggered when the player's character is beneath a certain water level threshold, at which point they transition into this state from either the on ground or in the air state. This update is intended to provide a more immersive and realistic player experience during gameplay. New swimming related events have also been added, such as onStartSwimming and onEndSwimming, to track and respond to the character's transition in- and out- of the swimming state. Appropriate event handlers have been inserted in the Animator Controller too, for visually representing the player's state.