Adjust Swimming Mechanics for Improved Buoyancy
Updated the swim state mechanics in the movements controller to improve the buoyancy when a pawn is in water. The original implementation was causing the pawn to sink too quickly. A temporary solution of adjusting the entity's velocity upwards has been applied to simulate buoyancy. We are multiplying Gravity with 0.3f and Time.Delta to achieve this. Further improvements are planned and documented in the TODO.