Add StaminaHandler and integrate it into MovementsController
Introduced a new StaminaHandler class which is responsible for managing the stamina level of a pawn. This system has been integrated into the MovementsController, ensuring that the pawn cannot sprint indefinitely, instead their speed will be reduced according to their current stamina level. This was achieved by using the StaminaHandler class in the MovementsController class and adding relevant transition conditions in the MovementsController.HFSM.cs. This change increases game realism by limiting player's movements to their stamina level. Also, it's important to note that the OnEnterSprintState now includes code to decrease stamina when sprinting and increase it otherwise.