Refactor variables related to stamina in game mechanics
Variable names related to stamina in game mechanics were confusing and not consistent. To simplify and make the code more understandable, `MaxStamina` was renamed to `MaxStaminaValue`, `CurrentStamina` to `Value`. These changes were reflected in the `MovementsController` class as well. Now, it should be straightforward to understand what these variables are for and how they are used in our game.