"Set max standable angle in movements controller"
This commit updates the MovementsController.OnGroundState.cs file to set the maximum standable angle in the movement helper. Before this change, the code didn't determine the maximum standable angle in the MoveHelper setup, potentially leading to incorrect movement behaviors. Now, the ground angle of the entity is considered, enhancing movement realism.