Add ability system and modifiers for health and stamina
This commit introduces an ability system with the addition of an IAbility interface, an AbilityController, and two kinds of modifiers: MaxHealthModifier and MaxStaminaModifier. This allows us to add different abilities to the game pawns and manipulate their max health and stamina. This paves the way for more flexible game dynamics. Also, the MovementsController has been updated to allow stamina to be overridden. The Pawn class has been modified to include max health.